+++ src/Kmd.Logic.Consumption.Client.AuditClient/AuditClientConsumptionMetricsDestination.cs
@@ -5,27 +5,45 @@ namespace Kmd.Logic.Consumption.Client.AuditClient
 {
     public class AuditClientConsumptionMetricsDestination : IConsumptionMetricsDestination
     {
-        private readonly IAudit audit;
+        public static string Template { get; } = "Consumed {Amount} for {Meter} on resource {ResourceId} in subscription {SubscriptionId}";
 
-        public AuditClientConsumptionMetricsDestination(IAudit audit)
+        public static string GetDefaultSubOwnerContextName(string propertyName) => $"__Sub_{propertyName}";
+
+        private readonly IAudit _audit;
+        private readonly Func<string, string> _getSubOwnerContextName;
+
+        public AuditClientConsumptionMetricsDestination(IAudit audit, Func<string, string> getSubOwnerContextName = null)
         {
-            this.audit = audit ?? throw new ArgumentNullException(nameof(audit));
+            this._audit = audit ?? throw new ArgumentNullException(nameof(audit));
+            this._getSubOwnerContextName = getSubOwnerContextName ?? GetDefaultSubOwnerContextName;
         }
 
         public IConsumptionMetricsDestination ForInternalContext(string propertyName, string value)
         {
-            return new AuditClientConsumptionMetricsDestination(this.audit.ForContext(propertyName, value));
+            return new AuditClientConsumptionMetricsDestination(
+                this._audit.ForContext(propertyName, value),
+                this._getSubOwnerContextName);
         }
 
         public IConsumptionMetricsDestination ForSubscriptionOwnerContext(string propertyName, string value)
         {
             return new AuditClientConsumptionMetricsDestination(
-                this.audit.ForContext($"__Sub_{propertyName}", value));
+                this._audit.ForContext(this._getSubOwnerContextName(propertyName), value),
+                this._getSubOwnerContextName);
         }
 
-        public void Write(string messageTemplate, params object[] propertyValues)
+        public void Write(Guid subscriptionId, Guid resourceId, string meter, int amount, string reason = null)
         {
-            this.audit.Write(messageTemplate, propertyValues);
+            var audit = reason == null
+                        ? this._audit
+                        : this._audit.ForContext("Reason", reason);
+
+            audit.Write(
+                Template,
+                amount,
+                meter,
+                resourceId,
+                subscriptionId);
         }
     }
 }