+++ CHANGES.md
@@ -1,2 +1,6 @@
+0.2.0
+* Fixed `resourceId` and `subscriptionId` being in the wrong message template positions
+* Improved `IConsumptionMetricsDestination` so we don't require message templates
+
 0.1.0
 * Initial version